openmp to find max min and mean
Assume that a program generates large quantities of a 
floating point data that is stored in an array. In order to determine 
the distribution of the data we can make a histogram. We simply 
divide the range of data into equal sized sub intervals of bins. 
Determine the no of measurements in each bin and plot a graph 
(bar) showing the relative sizes of the bin. Use MPI or OpenMP 
to implement the above and find the max, min and mean of the 
data.